% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpasbien.R
\name{extract_description}
\alias{extract_description}
\title{Extract description}
\usage{
extract_description(url)
}
\arguments{
\item{url}{cpasbien url where to extract the description from}
}
\description{
Extract description
}

