% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpasbien.R
\name{get_all_movies}
\alias{get_all_movies}
\title{Extract movies information}
\usage{
get_all_movies(pages = seq_len(n), n = npages("films"),
  .progress = "text", ...)
}
\arguments{
\item{pages}{pages to retrieve}

\item{n}{maximum number of pages}

\item{.progress}{see \code{\link[plyr]{llply}}}

\item{\dots}{further arguments for \code{\link[plyr]{llply}}}
}
\description{
Extract movies information
}

