% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpasbien.R
\name{search_episodes}
\alias{search_episodes}
\title{Search from series episodes}
\usage{
search_episodes(query, pages = seq(0, n), n = npages_search(query,
  "series"), ...)
}
\arguments{
\item{query}{what to search}

\item{pages}{which pages to retrieve}

\item{n}{total number of pages. The default values retrieves the total number of pages for this query}

\item{\dots}{further arguments to pass to \code{\link[plyr]{llply}}, e.g. \code{.progress}}
}
\description{
Search from series episodes
}

